refactor(scripts): rename to awto_riden_*, dedup, consolidate behind test/dev runners
Renames all importable scripts to the awto_riden_* convention (the two run-only entry points keep hyphens: awto-riden-test.py, awto-riden-dev.py). Updates every cross-reference: awto_riden.py's plot import, the report_pages/plot_waveforms imports, the timing-matrix subprocess path, dev.py's delegation map, and the active docs (README, REGENERATING.md, mr11_sine_test.md, copilot-instructions.md). Historical docs/reports manifests are left as-is (provenance records).

Dedup/merge: report_pages + backfill_reports_from_docs -> awto_riden_report.py (single module, backfill exposed as backfill_main()).

Deprecated/removed (now covered by awto-riden-test.py stages or folded in): transport_test.py (=status stage), output_test.py (=write stage + safe-state teardown), regenerate_timing_artifacts.sh (=awto_riden_timing.py).

Verified: 29 tests pass, all renamed modules import, both CLIs load, and gen-docs regenerates 10/10 figures end-to-end via the renamed modules.
