+++ README.md
@@ -27,7 +27,7 @@ Latest measured set summary:
 Regenerate all timing artifacts in one command:
 
 ```bash
-./scripts/regenerate_timing_artifacts.sh /dev/ttyUSB0 1.0 0.2
+python3 scripts/awto_riden_timing.py --port /dev/ttyUSB0 --voltage 1.0 --current 0.2 --mode both
 ```
 
 ## Quick start
@@ -169,7 +169,7 @@ Recommended one-command test set (includes fastest/no-cadence point `0 ms`):
 
 ```bash
 source .venv/bin/activate
-python3 scripts/timing_test_set.py \
+python3 scripts/awto_riden_timing.py \
   --port /dev/ttyUSB0 \
   --voltage 1.0 --current 0.2 \
   --mode both \
@@ -187,7 +187,7 @@ Direct matrix runner:
 
 ```bash
 source .venv/bin/activate
-python3 scripts/connected_load_timing_matrix.py \
+python3 scripts/awto_riden_timing_matrix.py \
   --port /dev/ttyUSB0 \
   --voltage 1.0 --current 0.2 \
   --poll-ms 0,20,50,100,150 \
